Abstract storyboard column component
